Django_coffee_house/requirements.txt
python-magic==0.4.27
python-utils==3.3.3
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-templated-mail==1.1.1
django-versatileimagefield==2.2
django-webpack-dev-server==1.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.1
Pillow==9.2.0
progressbar2==4.0.0
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
python-dotenv==0.21.0
python-magic==0.4.27
python-utils==3.3.3
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
django-filter==23.1
